[Feature] GameManager: complete shooting game
- 게임 상태를 액터 단위 로직과 분리해 GameManager로 관리
- std::function<void()> 기반 콜백으로 레벨 이벤트 연결
- 점수 획득과 플레이어 죽음 이벤트를 중앙에서 관리
- GameLevel이 게임 오버 대기 상태를 처리
- 게임 플레이 로직과 UI 표시를 분리하는 구조 적용

+++ ShootingGame/Actor/EnemySpawner.cpp
@@ -35,7 +35,7 @@ void EnemySpawner::Tick(float deltaTime)
 
 void EnemySpawner::Spawn()
 {
-    const uint64 count = enemyTypes.size();
+    const int32 count = (int32)enemyTypes.size();
     const int32 index = FMath::RandRange(0, count - 1);
     const int32 posY = FMath::RandRange(1, 10);
